+++ src/AiAltText.php
@@ -73,12 +73,6 @@ class AiAltText extends Plugin
         );
 
         $this->attachEventHandlers();
-
-        // Any code that creates an element query or loads Twig should be deferred until
-        // after Craft is fully initialized, to avoid conflicts with other plugins/modules
-        Craft::$app->onInit(function() {
-            // ...
-        });
     }
 
     private function attachEventHandlers(): void